Add option to enable sensors generation on devguide startup.
By setting SENSORS_GENERATION_ENABLED=true we can force the generation
of temperature and humidity sensors for each one of the restaurants
loaded on devguide when starting the container.  This may take a while
to complete, so it's disabled by default.  We can also force an initial
update of the sensors by setting SENSORS_FORCED_UPDATE_ENABLED=true.