Add tutorials.tourguide-app.config docker image.
This image will contain extra configuration for the official images
whithout having to use volumes from the host.  This will allow us to run
tourguide with docker-machine and deploy it on a remote server.  There
are still some problems that need to be solved to be able to deploy
tourguide with docker swarm.