Update the test scripts to use different container names.
As the container names are now forced via the container_name directive
in the compose file, the test scripts had to be updated to generate a
different name for the test containers, as this can't be done now via
the --project parameter of docker-compose.  Now, a temporary modified
version of the compose file is created and used with the required
container names.